+++ embodisteer/policies/jm2d.py
@@ -458,7 +458,6 @@ class DiffusionUnetTimmPolicyJM2D(DiffusionUnetTimmPolicyBaseline):
     def predict_action(
         self,
         obs_dict: Dict[str, torch.Tensor],
-        fixed_action_prefix: torch.Tensor = None,
         env_batched: bool = False,
         chunk_start_pose: torch.Tensor = None,
         obstacle_info=None,
@@ -485,12 +484,6 @@ class DiffusionUnetTimmPolicyJM2D(DiffusionUnetTimmPolicyBaseline):
             dtype=self.dtype,
         )
         cond_mask = torch.zeros_like(cond_data, dtype=torch.bool)
-        if fixed_action_prefix is not None and self.inpaint_fixed_action_prefix:
-            n_fixed_steps = fixed_action_prefix.shape[1]
-            cond_data[:, :n_fixed_steps] = fixed_action_prefix
-            cond_mask[:, :n_fixed_steps] = True
-            cond_data = self.normalizer["action"].normalize(cond_data)
-
         nsample, _ = self.conditional_sample(
             condition_data=cond_data,
             condition_mask=cond_mask,